backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
google-auth>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.13,<4.0
python-dotenv>=1.0,<2.0
drf-spectacular>=0.27.1
razorpay>=1.3.0
